# INTELLIVAULT-PROJECT-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IntelliVault AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Document management
  - name: auth
    description: Authentication
  - name: alerts
    description: Smart alerts and notifications
  - name: audit
    description: Audit trail logs
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocument"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List smart alerts
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit]
      summary: List audit trail logs
      responses:
        "200":
          description: List of audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"

  /stats:
    get:
      operationId: getStats
      tags: [documents]
      summary: Get vault statistics
      responses:
        "200":
          description: Vault stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [personal, education, financial, medical, legal, other]
        fileType:
          type: string
        fileSize:
          type: integer
        expiryDate:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        isEncrypted:
          type: boolean
        isSelfDestruct:
          type: boolean
        selfDestructAt:
          type: string
          nullable: true
        viewCount:
          type: integer
        maxViews:
          type: integer
          nullable: true
        fraudRisk:
          type: string
          enum: [low, medium, high]
        aiSummary:
          type: string
          nullable: true
        uploadedAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - fileType
        - fileSize
        - isEncrypted
        - isSelfDestruct
        - viewCount
        - fraudRisk
        - uploadedAt
        - updatedAt

    CreateDocument:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [personal, education, financial, medical, legal, other]
        fileType:
          type: string
        fileSize:
          type: integer
        expiryDate:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        isSelfDestruct:
          type: boolean
        selfDestructAt:
          type: string
          nullable: true
        maxViews:
          type: integer
          nullable: true
      required:
        - name
        - category
        - fileType
        - fileSize

    Alert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [expiry, fraud, access, security]
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        documentId:
          type: integer
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - type
        - title
        - message
        - severity
        - isRead
        - createdAt

    AuditLog:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        documentName:
          type: string
          nullable: true
        ipAddress:
          type: string
        location:
          type: string
        device:
          type: string
        timestamp:
          type: string
      required:
        - id
        - action
        - ipAddress
        - location
        - device
        - timestamp

    VaultStats:
      type: object
      properties:
        totalDocuments:
          type: integer
        encryptedDocuments:
          type: integer
        expiringDocuments:
          type: integer
        storageUsed:
          type: integer
        storageLimit:
          type: integer
        securityScore:
          type: integer
      required:
        - totalDocuments
        - encryptedDocuments
        - expiringDocuments
        - storageUsed
        - storageLimit
        - securityScore